requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
croniter==6.2.2
cryptography==46.0.7
distro==1.9.0
Django==6.0.3
django-rq==4.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.2
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.72.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
mutagen==1.47.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai-whisper @ git+https://github.com/openai/whisper.git@cba3768142a28276a90f14907e4900372c0c3ee0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
regex==2026.4.4
requests==2.33.1
rq==2.8.0
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
yt-dlp==2026.3.17
yt-dlp-ejs==0.8.0
